Fix critical issues

+++ src/main/resources/Init_db.sql
@@ -1,6 +1,6 @@
 CREATE TABLE books (
-    id INT NOT NULL AUTO_INCREMENT,
+    id BIGINT NOT NULL AUTO_INCREMENT,
     title VARCHAR(255),
-        price INT,
+    price DECIMAL(10, 2),
     PRIMARY KEY (id)
 );
